rename to ohsome_spec
